Update README.md&(0x7dB738F5c400f3734A9550aF79d0bF72339a6279)
I want my deposit to be deposited in foreign currency to my wallet, my country, even though I have paid all my crimes, of course, my judicial payments were made by the organization in 2021, unfortunately, I was still blocked from everywhere and I can't bear this oppression anymore. I plan the organizations and the basic foundations, but people are just abusing, of course, I am with those who think that they should eat garbage and be ashamed of the manger. I have to live in a fifty-meter house, and I am a tenant, and I don't even have the money for my rent. I can't see that you are oppressing and corrupting in all kinds of ways, and the person who was able to do the greatest work for the international community should be imprisoned, and you pressured me in all kinds of ways, shame on you.